+++ lib/src/horizontal_space_widget.dart
@@ -12,6 +12,8 @@ class HorizontalSpaceWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return SizedBox(width: width,);
+    return SizedBox(
+      width: width,
+    );
   }
 }